requirements.txt
python-dotenv~=0.21.0

Django~=3.2
django-phonenumber-field~=5.1.0
django-mptt~=0.12.0
django-extensions~=3.1.3
django-json-widget~=1.1.1
django-crispy-forms~=1.11.2
django-admin-sortable2==1.0
django-redis==5.2.0
django-sequences==2.6
django-solo==2.0.0
django-webpack-loader==1.6.0
django-log-request-id==2.0.0

# Django REST Framework
djangorestframework~=3.12.4
djangorestframework-jwt==1.11.0
drf-spectacular
drf-spectacular[sidecar]
django-cors-headers~=3.7.0
djangorestframework-simplejwt
drf-writable-nested~=0.6.2
drf-extra-fields~=3.4.0
django-filter~=2.4.0
markdown~=3.4.1
djoser~=2.1.0
django-rest-swagger~=2.2.0


# Utils
phonenumbers~=8.12.21
pyotp~=2.6.0
psycopg2-binary~=2.9.2
daphne==3.0.1
gunicorn==20.1.0
django-environ==0.9.0
channels==3.0.4
channels-redis==3.3.1
redis==4.1.0

# Celery
celery==5.2.3
flower==1.0.0
django-celery-beat==2.4.0
